Set default values for imageWidth and imageHeight
As @jorgefilipecosta mentioned,
this might help in cases where the warning still appears:

A component is changing an uncontrolled input of type number to be controlled. Input elements should not switch from uncontrolled to controlled (or vice versa). Decide between using a controlled or uncontrolled input element for the lifetime of the component.