Add check for author in post data before submission
Previously, the submission code on the Post Edit page assumed that the
Post data from the REST API contained an author field. But this is not
the case for CPT's that do not include `author` in their `supports`
array. This commit only includes the author in the submission if it
exists.