Fix broken active state on formatting buttons
At some point recently a small regression was introduced to the formatting button styles where the `:active` state box shadow that is inherited from the IconButton component lingered. To reproduce in master, click and drag out on a formatting button and note a gray box shadow lingering. This PR fixes that.

I'm unsure as to how the regression was introduced, but it was likely due to the rather long selectors the IconButton component uses, which would be good to refactor.