Add: note in the save function documentation to discourage side effects (#15765)
I noticed people are not aware side effects cannot be used in the save function, and people use the data module in the save function e.g: to query the post title. This may easily break the blocks.
This commit updates the documentation to make people more aware of this problem.