@wordpress/data: Introduce useSelect custom hook. (#15737)
A new `useSelect` hook for the `wp.data` api!  This hook allows for an alternative way to interact with data kept in a store state via registered selectors following a similar pattern to `useEffect` and other react hooks.

Other notable changes in this pull:

- exposing the custom `useRegistry` react hook for retrieving the `registry` object from components in a registry provider tree.
- `withSelect` implements `useSelect` under the hood.

Many thanks also to @epiqueras for contributions to the work in this pull.