Ensure icons for image/video are SVG components
In #15551, these were changed to support passing props, but that also means
that when you pass the video icon to the `Icon` component, it won't inject the
right size, since it's not a SVG component.

Instead, we can export the new function to allow for a customizable icon, while
exporting the SVG component by default.