Adds new generator based API signature to Block transforms
Previously the transforms function was pased two arguments

1. attributes
2. innerblocks

This wasn’t extensible and more advanced transformations require more information about the block (eg: name).

To avoid bloating the signature, a progressive enhancement approach is applied whereby if a generator function is passed as the transform then we pass the entire block object to the generator.

This is opt-in only and is backwards compatible with all existing transform functions.