Fix convert API to match established patterns for consistency
Previously `convert` always passed an array of block objects even if there was only a single block.

This is inconsistent with the implementation of the existing `transform` method which passes only a block’s attributes/innerBlocks pair when it is not a multi block.

To retain consistency with the existing `isMultiBlock` paradiagm this updates the API of `convert` to pass a single block object when not in multiblock mode.