Fix: DateTimePicker passes current seconds as the selected seconds value (#15495)
DateTimePicker does not allow users to pick a value for seconds, so when the user changes a time/date, it should look like a value of 0 was selected as the seconds value. Currently, the seconds of the current date are passed as the selected seconds, which may make it look like a "random" value was selected as the seconds value.

This causes a regression where WP_CRON missed some scheduled posts.