[RNMobile] keyboard focus issues (#15999)
* Focus RichText on mount if block is selected and says so

Some blocks have multiple RichText or a RichText among other children.
Example: Quote blocks has 2 RichTexts and Image block has a RichText for
the caption. We want to control when and which of those RichTexts will
request focus. On the web side, the DOM is used to search for a inputbox
to focus but on RN, we don't have a DOM to search.

Instead, this commit makes the assumption that a RichText always request
focus if its parent has passed `true` in `isSelected` and only if the
parent hasn't inhibited that behavior by using the `noFocusOnMount`
prop.

* Simplify the passing of noFocusOnMount

* Focus and selection logic closer to GB web's

* RichText's API uses unstableOnFocus at the moment

* Don't force selection update if position is unset

That can happen when the RichText based block gets tapped somewhere and
the caret position is not updated yet from Aztec.