Update i18n makepot scripts
These scripts stopped working because the gutenberg project changed to
only obtain translations from files in the build folder after the
project was built. As a result, gutenberg is not getting translations
for any *.native.js files. This addresses that by:

1. Performing a gutenberg build and obtaining a pot file from
`gutenberg/build/`.

2. Generating a gutenberg-mobile pot file based on the contents of
`./src/*.js` and `./gutenberg/*.native.js`, while also excluding
anything already contained in the gutenberg pot file that was
generated in step 1.