Adds default funcs as props
Previously the component assumed that the onChange and onInput props would always be present. However, there are times where this may not be the case such as when the BlockPreview component uses the provider to render the previews. In such cases `onChange` is called but is undefined which causes an error to be thrown.

To avoid this we provide lodash noop as defaults for both “change” props.