Adds check to determine whether event occured within element trapping focus
Previously there was no check to see whether the blur event occured within the wrapped component. If it occurs within then we do not want to trigger the `handleFocusOutside` callback because (by definition) the focus has not moved outside the wrapped element.

This may have been a regression introduced in https://github.com/WordPress/gutenberg/pull/16878