Add disable support to PostStatus settings panel
While many other panels support the isEditorPanelEnabled option, the
PostStatus panel does not. If the 'post-status' panel is not enabled, the
component now returns null. Previously, it always returned the full
Panel regardless of whether it was enabled.

This is cleaner than returning null, since ifCondition does that for us.