Try: Fix MacOS auto-hiding scrollbars disappearing.
DO NOT MERGE. This is a work in progress and needs discussion first.

This PR fixes #16966.

However it does so by some weird Voodoo. For whatever reason, the clip-path attached to screen-reader-text is causing the silly MacOS auto-hiding scrollbars to disappear entirely. But only in Chrome (and Chromium based Edge), and only when there are adjacent elements with specific styles.