Try: Fix cropped focus ring in block library
Maybe fixes #17094 and maybe fixes #17168 too.

There's focus ring on blocks in the block library and the switcher. These are cropped when blocks are near the edges, due to an overflow: hidden;. This PR attempts to fix that with a deadpan negative margin. Will it work? This is where YOU come in!