Add disable support to PostStatus settings panel (#17117)
* Add disable support to PostStatus settings panel

While many other panels support the isEditorPanelEnabled option, the
PostStatus panel does not. If the 'post-status' panel is not enabled, the
component now returns null. Previously, it always returned the full
Panel regardless of whether it was enabled.

This is cleaner than returning null, since ifCondition does that for us.

* Use isEditorPanelRemoved instead of isEditorPanelEnabled

We do this since the panel should never be disabled with the enabled
selector, but it can be programmatically removed.

* Update logic for removing post status

* Update e2e test to remove PostStatus panel

* Update e2e test to check if Status Panel exists