Try: Show horizontal movers for Social Links.
This is more of a proof of concept and experiment, than it is final PR. So *do not merge as is*.

We might find that we like this, in which case it might need a little polish, and _then_ we can ship it. But for now this PR has been created to help further discussion in #16637.

This PR does the following:

- Starts with Social Links, because that is the first block to "absorb child block UI", and therefore a good test-bed for this.
- Additionally, Social Links has been merged with _no mover controls_. This PR resurfaces them.
- It restyles, positions, and rotates the mover control to be horizontal.
- It restyles and positions the ellipsis menu.

The overall pattern it mimics, is that of gallery items, which also have horizontal movers, even though those are not yet technically child blocks.

Your feedback is much apprecated.