LocalAutosaveMonitor: Use immediately available autosave data
* Until now, although we were requesting the entire autosave entity
inside `LocalAutosaveMonitor`'s `useAutosaveNotice`, we only care about
the existence or absence of an autosave for the current post.

* `useAutosaveNotice` was requesting the entire entity because there
were plans in #17501 to compare local and remote autosave objects, a
requirement which was later dropped.

* In contrast, the editor setting `autosave` is immediately available
upon editor initialisation.

* Later on, _if_ we choose to reconcile remote and local autosaves with
more sophistication, this could possibly be done based on recency of
autosave -- in which case we could might still get away with editor
settings if the server provides this information inside the `autosave`
object.

* This change should hopefully provide a more reliable page-loading
experience for autosave's E2E tests.