update tests due to new expectations.
With the fix implemented in the tested hook, expecations for initial state of the hook changes since the effect bails early when the current state of `isSmall` from the viewport store matches the previous state stored on the ref.

On initial render, this hook will always bail early.

I also added some inline comments to help explain what is happening in the test and what is expected.