Distinguish user-input from structural HTML text
In this patch I'm starting to propose some shifts that make a more
intentional distinction between text that someone _typed_ in to the
editor and text that a component produces while rendering to HTML.

A motivation for this change comes from the code editor, or a paragraph
block where someone enters valid HTML markup. We want `<hr />`, when
typed in, to _always_ appear as `<hr />` when viewing the rendered page.

It happens that a combination of transformations in the editor and also
within the WordPress core can unintentionally mangle these user-input
sequences. Sometimes it's because WordPress is trying to preserve them;
sometimes it's because we're trying to filter-out unwanted input.

In this patch I've created new `escapeHTML` and `unescapeHTML` functions
whose roles are to serialize and unserialize all user input so that we
can prevent that mangling later on. Currently they are implemented by
replacing characters with their matching named character entities. In
the future it may not stay this way.

This patch is incomplete but it's meant to start the journey.