Polish search widget. (#17648)
This applies a max-width.

Fixes #17593 — the other half of it.

Basically the search widget can grow as wide as your main body content column will let it. This PR applies a max-width.

Do we want this? If you insert the search widget in a really wide space, then it will grow to fill it. Not sure if good or bad, but it can definitely look a little overwhelming.