Full-height columns fix in IE11.
This PR first of all makes the columns block usable in IE11. It was barely functional before.

Secondly, but I'm not entirely sure how to test this, it appears to disable full-height columns previews, but I'm not entirely sure exactly what that means. It does so because the rules to enable what that is, break the entire block for IE11, so they have been wrapped in a "supports".

Depending on what it is, this is probably fine, and we can see that preview as a progressive enhancement.