Fix checkboxes for postmeta.
This fixes #18053.

It is not a great fix, but it is nonetheless the least disruptive I could think of, and solves the problem now.

The better, longer term, fix is to roll out the singular checkbox and radio button style from the block editor to all of WordPress, in which case the code written for this PR can simply be removed entirely.