Refactor ActionBar with divider + support to (maybe) render Toolbar
This update refactors how the styles are defined within ActionBar and
it's sub-components. This allows for greater flexibility, and can be more
easily combined with something like a `Divider`.

An experimental `Toolbar` component was added in the Story to see test
how compatible the `ActionBar` API with a Toolbar like interface.