Remove iframe and img responsiveness rules.
This is an alternative, and/or addition to #18240.

There is discussion around removing these base styles in favor of letting the theme do the work, and the img and iframe rules have been around for a while but might not be needed. This PR removes both those rules, the other only the iframe which seems to be the most offending.