Updates implementation to remove state and utilise dispatch to update innerBlocks
After a long discussion with @mtias we realised that unlike other similar implementations (Media+Text and Columns) of a placeholder, this one couldn’t use the “Template Options” feature of <InnerBlocks> to achieve the custom design required. Instead it rendered a `<Placeholder />` directly based on conditions.

As a result, we now use the `replaceInnerBlocks` from `block-editor` to dispatch an update to innerBlocks manually when the placeholder buttons are clicked.

Addresses

* https://github.com/WordPress/gutenberg/pull/18363#discussion_r343694455
* https://github.com/WordPress/gutenberg/pull/18363#discussion_r343833490
* https://github.com/WordPress/gutenberg/pull/18363#discussion_r344013908