Components: ResizableBox, add showHandle prop (#18097)
* Components: ResizableBox, Add showHandle prop

This update adds a new prop to the ResizableBox component. This new prop
`showHandle` (`bool`) determines if the draggable resize handles should be
visible or not.

* Add ResizableBox storybook example

This update adds a default story for ResizableBox. Knobs have been added
to the story to better demonstrate the effects of the props.

* Rename is-show-handle to has-show-handle

Refactor styles to consolidate selected and show handle

* Refactor ResizableBox story example

Moves the inline-styled div to the parent. Allows for the Example
component to read more clearly.