Adds in a styled drop down option
It should be discussed if this is by default or not for subnav. It could be a style or it could be one we add by default.

I have styled it dark but we could also go with white as an option, which might work better across all text colors.