Allow parent Block to consume child Block's toolbar (#18440)
* Adds primitive hardcoded child toolbar appearing in parent toolbar area

* Update to dynamically test for true block support for the “consume” toolbar feature

* Update implementation to utilise InnerBlocks prop

Previously Block settings controlled whethere the toolbar was “consumed”. Now this is controlled by `InnerBlocks`.

* Adds classname to signify consumption of toolbars

* Fix event bubbling via use of Slot Fill

Still a primitive PoC.

* Add comments for clarity

* Remove unusued isRootOfHierarchy prop

* Rename var to improve clarity

A parent of X denotes a the immediate progenitor of X. Using `hasSelectedInnerBlock` with the `deep` argument checks all levels of nesting of InnerBlocks. Therefore the value could be the parent but it could also be a grand-parent (and so on). As a result a more useful term is “ancestor”.

* Update to only show toolbar on immediate parent

This was happening anyway because Slots can only be rendered once. But this code makes it clearer as to the intent.

* Updates to allow deeply nested toolbars to display toolbars on root ancestor

* Update prop to better reflect usage

We’re not consuming toolbars as Block. We’re telling InnerBlocks to capture. The parent Block itself has no setting to denote that it is consuming toolbars.

* Adds optional BlockTitle to BlockToolbar

* Display current Block name within “captured” child Block toolbars

* Commit prop name change on Nav Block

* Adds select to get blockListSettings for a set of Blocks (given clientIds)

* Updates to implement captureChildToolbar with new “parent only” API

See https://github.com/WordPress/gutenberg/pull/18440#issuecomment-554621877

* Update to implement new dual API for capturing

There are now two props available which handle capturing differently.

See https://github.com/WordPress/gutenberg/pull/18440#issuecomment-554621877

* Update selector docs

* Documents new capture API on InnerBlock props

* Reinstate lost changes during rebase.

* Updates to cache expensive selector

* Updates to move childtoolbar Slot/Fill outside of render.

* Updates to move ChildToolBar slot to separate file

* Fix Toolbar display logic to account for forced toolbars

Addresses https://github.com/WordPress/gutenberg/pull/18440#discussion_r348085860

* Revert "Display current Block name within “captured” child Block toolbars"

This reverts commit f40e8f2bf57eb1e4ce20013b688e28a109021e6d.

Removing because it muddies the PR and starts to address more complex UI issues. Limiting the PR scope by reverting this.

* Update to DRY up usage of BlockContextualToolbar

* Fix issue with forcing toolbars and refs due to rendering x2 toolbars

Previously we were rendering x2 toolbars which cause the ref associated with forcing toolbars to become out of sync. This meant that using keyboard shortcuts to focus the toolbar didn’t work. Fixed by rendering component on demand.

* Fix linting issues

* Simplify to single API for deeply capturing all toolbars.

* Removes capturing toolbars on Nav Block

As the API is experimental, remove it from any public Blocks for now.

* Fix to reuse existing var to DRY up conditionals

Addresses https://github.com/WordPress/gutenberg/pull/18440#discussion_r352724549

* Remove selector params from createSelector usage

Addresses https://github.com/WordPress/gutenberg/pull/18440#discussion_r352725751

* Add classname to block to indicate that toolbar is captured

* Update focus/selected visual indicator to take account of missing toolbar

* Fix classname to indicate captured toolbars to apply without selected condition

* Reinstate changes lost during complex rebase