Update README.md
I've made some minor grammatical changes, removed some wordy bits or bits that didn't seem to further the overall message. Updated the phrasing to note that Phase 1 is not future but saw its initial iteration in 5.0 and removed in 2019 for Phase 2, since this doesn't seem to be accurate (but I'm unsure of the updated timetable).