Components: Dashicon: Improving mapping and Stories
This update introduces a key/value object map for the support Dashicon types.
Originally, the Dashicon Paths were defined and targeted with a large `switch`
statement. Changing it to a key/value map enables the collection to be accessed
outside of the component. This is useful for cases like the DashIcon stories,
allowing us to loop through all of the entries to render examples for each
icon.