Editor: Fix Block Embed Input size (#19438)
This update fixes the Input vs. Button sizing and alignment as seen in Embed
blocks, such as Instagram, Twitter, or YouTube.

The problem stemmed from the Placeholder component from `@wordpress/components`.
The `display: flex` container caused the `input` to grow responsively
(height wise), which was triggered by the mysterious margin below the button.

The solution involved neutralizing the margins for both the `input` and `button`
elements, allowing for their heights and alignments to match.