Block: Fix form element alignment for Embed + Table
This update fixes the Input x Button alignment for both the Embed and
Table blocks via the internal Placeholder component.

The issue was caused by a recent update made to the base Placeholder
component which normalized it's margins. However, due to the element/layout
rendering of the Table block, it required certain margin values for the
Placeholder button to align correctly.

The solution was to revert the globalized changes, and apply the normalization
specifically to the Embed block.

I don't think this will be a permanent solution given how unpredictable
component combinations may affect layout (via margin). For now, this
solution is adequete to resolve these 2 Blocks.