Fix to allow keyboard arrow to move focus into suggestions when present
Previously it was not possible to use the down arrow key to move the focus into the search suggestions when initialSuggestions were displayed. This was due to requirement for there to be a value in the `input` which is now no longer valid now that initial suggestions can be displayed.

Note that removing the derived setting of selectedSuggestion state is valid as selectedSuggestions are always reset on each new data fetch anyway.

Resolves issue noted in https://github.com/WordPress/gutenberg/pull/19458#issuecomment-571789573