Wrap color palette in fieldset with label inside of a legend (#19546)
* Wrapped color palette in fieldset with label inside of a legend

Color palette popovers were not having their label read on screen readers, because they were just a div and span. This commit wrapps the toggle buttons in a fieldset with the label as a legend so it will read the label wwhen the component receives focus.

* Updated snapshot for color-palette component

The markup for the color-palette was changed to add a fieldset and legend, so the snapshot needed to be updated for tests to pass.

* Fixed color button selector paths in e2e tests.

Also updated the snapshots, as I changed the colors to be more obviously changed when watching the tests running in the browser.