Core-data: more granular cache invalidation when updating records
Currently, if you modify a single record, the cache is invalidated for all the records.

If the user is interested in a lot of entry records, this behavior can be very expensive.  In my case, I have 300 candidate objects stored as Entity Records and they're worth 800KB in JSON. If you modify. say. a single candidate's region data, the whole cache is invalidated and all the 300 candidates are re-fetched from the server. This commit separates the two caches of (`getEntityRecords` vs `getEntityRecord`). And such, flushing `getEntityRecord(X, Y, Z)` only invalidates the cache for a single record `X, Y, Z`. And all the other records remain intact. Much more efficient, at least in my case.