BaseControl: Rendering as a fieldset + RadioControl
This update enhances the `BaseControl` component with the ability to render
as a `fieldset` DOM element. This is necessary for correct semantics + better
a11y support for use cases with multiple inner fields. An example would
be a collection of `radio` buttons.

We have a component, `RadioControl`, that falls under this exact use case.

With this update, we can adjust the inner `BaseControl` of `RadioControl` to
render as `fieldset`.