Social Link: Remove unnecessary block deprecation
This PR introduces `core/social-link` as a new block type. Strictly
speaking, any concerns of backwards compatibility should only deal with
the migration of `core/social-link-FOO` types to this new one --
something that the parser is handling directly, populating the new
`service` attribute in that process.

Thus, no specific provisions are needed for the extant but obsolete
`site` attribute: the attribute will be silently dropped during block
validation. This can be attested by loading any content with links in
the old `core/social-link-FOO` shape.