Schedule to request focus instead of asking immediately
My current theory is that the View is not yet in the layout and
requesting focus to it causes a scroll to the top of the ScrollView,
since the coordinates of the view are not yet updated in respect to its
parent. I can be wrong, haven't fully reached the bottom of this issue.