Block Library: Avoid column width auto-adjustment when sibling width changes (#19515)
* Block Library: Avoid column auto-adjustment when width changes

* Block Library: Column: Add step attribute to column width range input

Without this attribute value, if a user would enter a decimal value in the input range, the result of Element#checkValidity (in RangeControl) would be `false`, thus making it difficult/impossible to enter non-whole-number widths (e.g. `33.3`).

There's a balance here between the effects of step, where ArrowUp/ArrowDown will now increment/decrement in smaller intervals than previously. This becomes further problematic if we want to support smaller decimal values (e.g. 33.33). A potential alternative would be to change the validation behavior of RangeControl to use a custom validation procedure in place of `Element#checkValidity`.

* Edit Post: Remove horizontal margins from notices rendered in sidebar

* Block Library: Column: Show Notice when total width is not 100%

* Block Library: Show Columns block range control only when "Manual" width

* Revert "Block Library: Show Columns block range control only when "Manual" width"

This reverts commit 92b8fd4bf2c7bbf57189095f02dfda8798667261.

* Revert "Block Library: Column: Show Notice when total width is not 100%"

This reverts commit e25b28b5e1393642854888b715c18eb31382aa2f.

* Revert "Edit Post: Remove horizontal margins from notices rendered in sidebar"

This reverts commit 9f79462713723712257c9710953cff76e463f123.

* Block Library: Column: Add placeholder label for auto width

* Block Library: Columns: Consider explicit only if all blocks

Avoid situation where adding or removing a Column in a Columns would forcefully assign width to a block which did not previously have a width. Only opt to redistribute widths to explicit numbers if all blocks already have widths.

* Block Library: Column: Grow to maximally fill space