Ensure Create button is “in view” by removing other results.
It’s possible that on certain screen sizes there is not enough room to display the Create button without having to scroll the LinkContorl seaerch results panel. This could cause the selection of the button to fail.

Testing if this fixes the broken e2e tests or at least makes them more resilient.