Fix to not render create button if there is no search term.
This was a throw back to a previous state of the `LinkControl` component whereby we wanted to render a create option to allow the user to create blank pages. This was removed as a requirement but the component was not fixed to account for that.

Addresses: https://github.com/WordPress/gutenberg/pull/19775#discussion_r379590227