RangeControl: Improve initial hover interaction with Tooltip (#20219)
* RangeControl: Improve initial hover interaction with Tooltip

This update improves the initial hover/mouseenter experience with the RangeControl's Tooltip. This is achieved be debouncing the interaction. By debouncing, the Tooltip rendering is less jarring when there are multiple `RangeControl` components next to each other.

A Storybook story was added to test and simulate this experience.

Lastly, the utility functions/hooks from `RangeControl` was abstracted to a dedicated `utils.js` file under the component directory.

* Simplifies setValue to be a plain function vs using useCallback

* Remove local ref from useCallback hook

* Remove guard for clearTimeout

* Add clearTimeout callback on unmount

* Improve Tooltip rendering for current RangeControl changes...

...with other RangeControl hover interactions.

* Removed whitespace + adjusted setState naming