Fix block selection indicator.
The behavior of manipulating blocks in the recent G2 refresh regressed slightly in #19701, this PR addresses that.

Specifically, the blue border is meant to indicate _block focus_, i.e. when it is present you can press Delete to remove the block. It is not meant to indicate the block boundaries when focus is inside text.