Fix auto-hiding appender regression.
In #20753 I caused a regression that collapsed the big square appender that exists in an empty columns block, or an empty group block.

The only thing that should collapse, until selected, is the black appender button inside blocks like Social Links, Buttons, or Navigation Menu.

This PR fixes that regression by scoping the animation to said button.