Compat: Use `empty` to detect imageSizes
`is_array` still warns if not set. Coincidentally helps avoid an unnecessary call to wp_get_registered_image_subsizes when there's no sizes to work with. One difference is that imageDimensions would not be set unless imageSizes is non-empty. But if the idea is for imageDimensions to be dependent upon imageSizes here, it's a sensible caveat.