Env: Bind "core" files to tests environment
When using a local source for "core" we want to ensure two things:

1. That changes the user makes within the "core" directory are served in
   both the development and tests environments.
2. That the development and tests environment use separate databases and
   `wp-content/uploads`.

Only (2) was working because the tests environment served everything
from a totally separate copy of WordPress.

The fix is to bind every file except wp-config.php and wp-content from
the local "core" directory to the tests environment.