Update the "patterns" to reflect "user interface" instead
This file was confusing because it used terminology like "patterns" for specific UI sections. We are using the terminology "patterns" in relation to Block Patterns nowadays, so this needed updating. I also changed the Title of the doc, so this may need to be reflected elsewhere?

There were several text changes made to align with how the current interface behaves. 

I did not update images to reflect the new G2 design yet. That should come in another PR.