Fix the JSDoc return type for getEntityRecords
`getEntityRecords` can actually return null based on the return type
of `getQueriedItems` which it calls. I can confirm it behaves this
way in the implementation so its a bug with JSDoc type for
getEntityRecords.

I did wonder if this would be desired behaviour though. Since
getEntityRecords is a selector I think it could make more sense
for it to always return an empty array to represent its empty state.