Core Data: Improve ifNotResolved typeability
Not yet able to configure this file to be type-checked, because `@wordpress/data-controls` does not yet have types.

Also, these changes aren't strictly ideal because the purpose of `@template` was to indicate that the signature of the input function would be respected for the "output" function (as a higher-order function).