Gallery block: Make gallery-image figure focusable
Currently, when you try to move images around
the gallery block, the arrow buttons fail in Safari,
this is because the arrow icons are unfocusable,
i.e having (`focusable=false`). Meaning a blur event
is fired when you click on the arrow icons, this blur
event disables the buttons thus canceling the click event.
This diagnosis is a bit speculative because and I'm not 100%
confident about it. But it holds true for my suggested fix here.

Fixes: https://github.com/Automattic/wp-calypso/issues/40531