Use the actual object's handles rather than duplicating its initialization
We were duplicating WP_Scripts and WP_Styles initialization
and therefore misrepresnting what was handles were actually
accessible through this API. By using the object, which for both
these classes is accessed through a getter for the static global
for each WP_Dependencies subclass, we can return the handles
that are actually available. This makes it so that any script
added by other plugins will also appear, for example.

This doesn't change the accessibility of those handles, they were
always already accessible.

Something to consider is whether we should implement a blacklist,
if there are scripts/styles that must never be requested in this
way, as in, if they were then we're doing something really wrong.
I'm not sure if any scripts like that exist today.