Use dependencies REST API to fetch style and script URIs
Rather than using `load-scripts.php` which only loads a small
subset of dependencies into a locally managed WP_Scripts/WP_Styles,
we can use the two new REST endpoints that use the globals
available throughout the rest of WP.

Each endpoint returns the full list of the URIs for the requested
dependencies as well as their respective dependencies. When multiple
dependencies are passed as in this format:

	/wp/v2/scripts/?dependency=dep_1&dependency=dep_2

The endpoint will return:

	[
		{
			handle: 'dep_1-dependency',
			src: '...',
		},
		{
			handle: 'dep_1-actual',
			src: '...'
		},
		{
			handle: 'dep_2-dependency',
			src: '...',
		},
		{
			handle: 'dep_2-actual',
			src: '...',
		}
	]

That is, it will return descriptions for the list of
dependencies you pass to the endpoint, as well as those
dependencies dependencies, in the correct order in which
they would need to be loaded.

That means we can consume the response and just create a
script or link element for each response to be able to
load the dependencies we've requested.

We'll also store the actually loaded dependency handles
rather than the ones requested through the props. The thought
here being that one component might have a dependency
which itself depends on the dependency of another component.

Using the example response above, if component Foo depends on
`dep_1-dependency` and Bar depends on `dep_1-actual`, if Bar
is rendered first, then Foo does not need to have its dependencies
asynchronously loaded as they were already loaded as part of
loading Bar's dependencies.

At this point, this code doesn't actually fully work for
the classic block. Something is still broken and I'm still
trying to work out what exactly it is. However, the scripts
and styles do indeed load and the initialization is successful.

I _think_ there might just be a problem where something isn't
being correctly awaited, but I've run out of steam today to
figure out just what's going on.