Guarantee script execution order
See inline comment for explanation for why this is necessary.

Also refactors the dependency loading code into a single function.
All of the logic is shared except for the actual DOM element creation.
While the component was pretty simple to understand on one read
through when there was no need for blocking, it didn't make sense to
repeat that complexity and the comment for it.

This also lets the component hide the complexity of loadScripts and
loadStyles from the exported function, as well as the already-loaded-
dependency tracking inside those closures.