Fetch TinyMCE translations from REST API
Uses the "private" _WP_Editors class to retrieve the TinyMCE
translations JSON blob and requisite metadata for initializing
the editor on the frontend.

If we decide to follow this approach, we'll want to refactor
the methods used from _WP_Editors into free functions, or else
move them into a separate class that is only used for getting
the TinyMCE translations.

One marked improvement we could achieve is to avoid json_encoding
the translations only to have the JSON.stringified on the client.
We should just send them down as an object rather than an encoded
string. I didn't do that for this commit because it's going to
require a wee bit of refactoring in wp_mce_translation to get
the raw translation object back rather than a string. That refactoring
should just go along with the refactor that pulls stuff out of
_WP_Editors.