No need to bind with current this
Since the functions don't use this, we can safely use null. Also helps
with type-checking :-)

We still need the bind, even if it's not a class, because of `settings`.
Not sure if syntax like `text => stripTags( settings, text )` will be
more readable, considered it, but decided to liimt the surface of my
changes.