Unset the inherit for links.
Mitigates and closes #22139.

The better fix would be to not set any styles for the editing canvas, this can be fixed if we are able to bring the editing canvas into an iframe.

In the mean time, this removes the "inherit" property. That means link colors are now the same color as links in wp-admin because that color bleeds in. While the bleed is undesirable, I was  unable to find a different fix that enables the vanilla editor styles to have blue links. Both `inherit`, `initial` and `unset` appear to keep the link color black. The only other fix would be to find a new blue color, and set it here, which would simply add specificity.