Insert block in document, not template part
A change to the inserter was causing the edit meant
for the document root to be applied to the template
part instead. Obviousy, this meant that the test
considered the wrong entity to be dirty, which resulted
in a failure. This updates the test to always insert the
block in the document, and not in a template part.