wp-env: add an interactive bash command (#22611)
Resolves #21301. The run command now supports more interactive
commands like bash or wp-cli shell without any extra options.
One benefit of this approach is that since stdio is inherited
from the calling terminal, information should be logged to the
terminal immediately without waiting for the process to finish.
This means some long running commands (like test-php) give more
immediate feedback.

This was implemented by spawning docker-compose on the CLI
directly instead of calling the npm wrapper for it. The npm wrapper
does not really support the interactive behavior we want.