Remove key UP/DOWN + SHIFT tests for NumberControl and InputControl
The reason is because we're now relying on the native HTML `input[type="number"]` incrementing/decrementing function.
Jump stepping values (holding shift to jump by a multiplier, default of 10) is now being handled by the new useJumpStep hook. This hook listens to `Shift` keyboard presses and modifies the `step` prop that is passed into the `input[type="number"]` element.