Refactor button icon logic
Previously, to render the default icon, we were passing an SVG through the icon prop to the @wordpress/components Button. The data provided for user uploaded site icons, however, are incompatible. We receive custom URLs, and not SVGs, for custom site icons.

Rather than support the logic for both Button icon props and site icon image tags, we do away with the Button icon prop entirely.

We now explicitly define the children for Button, which consolidates the display logic for both default and custom site icons.